add geoip processor, requiring local maxmind db
The initial support for the `geoip` processor has access to neither the
databases that are a part of the Elasticsearch cluster nor those that are
maintained by Logstash Core's GeoIp Database Manager.

While we plan to add support for using managed databases from Logstash Core's
GeoIp Database Manager (which will obviate the need to provide those managed
databases), until this is added any configuration that wishes to use Elastic
integrations that include GeoIp processors should be configured with the
path to local-on-disk directory containing the necessary Maxmind
databases that are typically part of an Elasticsearch cluster and are
referenced by name in those integrations:

 - GeoLite2-ASN.mmdb
 - GeoLite2-City.mmdb
 - GeoLite2-Country.mmdb

Forward note: when managed database support is added, databases found via
this explicit setting must _override_ same-named databases that are available
from the database manager in order to be backward-compatible with current use
of this plugin.