fix: ensure JVM 17 before loading jars & dependencies
While the default distribution of Logstash includes a Java 17 runtime, we also
support a bring-your-own-Java path and only require Java 11 to run Logstash
itself.

Here we refuse to instantiate the plugin on unsupported JVM's _before_ loading
the vendored jars, which prevents an opaque and eventual failure that could
occur later in processing when a classfile that was compiled for JDK17+ is
eventually loaded (e.g., when an Ingest Pipeline is first executed).