fix: handle IngestDocument metadata fields separately
The ES-ingest reserved Metadata fields by definition _cannot_ be top-level
fields of the events we send to Elasticsearch and therefore should not be
included in their reserved place in the Event that we generate from an
IngestDocument.

We already safely extract the `_version` from the IngestDocument's metadata
and safely map it to our LS-reserved `@version`, but there is reason to
believe that _any_ of these fields could be directly useful to a Logstash
pipeline that will eventually persist the events to Elasticsearch.

Here, we make all of these fields, sans their underscore prefix, available
in the metadata-key `[@metadata][_ingest_document]`, which prevents them
from clashing or being included in output accidentally but makes them
available to the downstream LS pipeline if the author wishes to access them.