fix: preemptively send basic auth
When Elasticsearch is configured to allow anonymous auth, it does not include
a challenge header in response to requests, so we cannot rely on the apache
http client's auth handling (which is overkill anyway, since an individual
Elasticsearch client is not connected to multiple realms).

Instead, we build the http basic auth header ourselves, and register an
interceptor to ensure it is added to all requests, similar to how we handle
api key auth headers.